requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
scikit-learn>=1.5.0
xgboost>=2.0.3
numpy>=1.26.4,<3.0.0
pandas>=2.2.2
openpyxl>=3.1.2            # required for reading .xlsx files

# ── MLflow ────────────────────────────────────────────────────────────────────
mlflow>=2.13.0

# ── DVC ───────────────────────────────────────────────────────────────────────
dvc>=3.50.0
dvc-gdrive>=2.20.0         # optional: for Google Drive remote

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.1
pyyaml>=6.0.1
prometheus-client>=0.20.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
httpx>=0.27.0
pytest-asyncio>=0.23.6

# ── Code quality ──────────────────────────────────────────────────────────────
ruff>=0.4.4